SP5: separate cookbook test project
