+++ src/ScottPlot5/ScottPlot Cookbook/IRecipe.cs
@@ -1,9 +1,8 @@
-﻿namespace ScottPlot_Tests.Cookbook;
+﻿namespace ScottPlotCookbook;
 
 public interface IRecipe
 {
     public string Title { get; }
     public string Description { get; }
-    public Category Category { get; }
     public void Recipe(Plot plot);
 }
