+++ src/ScottPlot5/ScottPlot Cookbook/RecipeGroup.cs
@@ -1,0 +1,7 @@
+﻿namespace ScottPlotCookbook;
+
+internal abstract class RecipeGroup
+{
+    internal abstract string Name { get; }
+    internal abstract string Description { get; }
+}
