+++ src/ScottPlot5/ScottPlot Cookbook/Usings.cs
@@ -1,0 +1,2 @@
+﻿global using NUnit.Framework;
+global using ScottPlot;
