--- src/ScottPlot5/ScottPlot5 Tests/Cookbook/Category.cs
@@ -1,7 +1,0 @@
-﻿namespace ScottPlot_Tests.Cookbook;
-
-public struct Category
-{
-    public string Name { get; set; }
-    public string Description { get; set; }
-}
