--- src/ScottPlot5/ScottPlot5 Tests/Cookbook/CookbookRecipes.cs
@@ -1,12 +1,0 @@
-﻿using ScottPlot_Tests.Cookbook.Recipes;
-
-namespace ScottPlot_Tests.Cookbook;
-
-public static class CookbookRecipes
-{
-    public static IRecipe[] GetAllRecipes()
-    {
-        // TODO: use reflection to return this
-        return Array.Empty<IRecipe>();
-    }
-}
