+++ src/ScottPlot5/ScottPlot5 Tests/GenerateTests.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot_Tests.RenderTests.Figure;
+﻿namespace ScottPlotTests.RenderTests.Figure;
 
 public class Tests
 {
