+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/Figure/BasicImageTests.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot_Tests.RenderTests.Figure;
+﻿namespace ScottPlotTests.RenderTests.Figure;
 
 internal class BasicImageTests
 {
