+++ src/ScottPlot5/ScottPlot5 Tests/RenderTests/Plottable/PieTests.cs
@@ -1,4 +1,4 @@
-﻿namespace ScottPlot_Tests.RenderTests.Plottable;
+﻿namespace ScottPlotTests.RenderTests.Plottable;
 
 internal class PieTests
 {
